Update global SDK version Organized unit tests
