+++ global.json
@@ -1,5 +1,6 @@
 {
+  "projects": ["src", "test"],
   "sdk": {
-    "version": "5.0.302"
+    "version": "5.0.400"
   }
 }
