+++ test/TextExtensionsTests/CleanExtraLineBreakTests.cs
@@ -1,12 +1,13 @@
 ﻿using FluentAssertions;
 using System;
 using Xunit;
-namespace PowerUtils.Text.Tests
+
+namespace PowerUtils.Text.Tests.TextExtensionsTests
 {
+    [Trait("Extension", "CleanExtraLineBreak")]
     public class CleanExtraLineBreakTests
     {
         [Fact(DisplayName = "String only empty should return empty")]
-        [Trait("Extension", "CleanExtraLineBreak")]
         public void CleanExtraLineBreak_Null_ReturnNull()
         {
             // Arrange
@@ -23,7 +24,6 @@ namespace PowerUtils.Text.Tests
         }
 
         [Fact(DisplayName = "String only empty should return empty")]
-        [Trait("Extension", "CleanExtraLineBreak")]
         public void CleanExtraLineBreak_Empty_ReturnEmpty()
         {
             // Arrange
@@ -40,7 +40,6 @@ namespace PowerUtils.Text.Tests
         }
 
         [Fact(DisplayName = "String with double spaces")]
-        [Trait("Extension", "CleanExtraLineBreak")]
         public void CleanExtraLineBreak_WithDoubleSpaces_resultEqualsInput()
         {
             // Arrange
@@ -57,7 +56,6 @@ namespace PowerUtils.Text.Tests
         }
 
         [Fact(DisplayName = "String with line breaks must keep the double spaces")]
-        [Trait("Extension", "CleanExtraLineBreak")]
         public void CleanExtraLineBreak_LineBreaks_ReturnStringWithLineBreaks()
         {
             // Arrange
@@ -74,7 +72,6 @@ namespace PowerUtils.Text.Tests
         }
 
         [Fact(DisplayName = "String with double line breaks - Hello world test")]
-        [Trait("Extension", "CleanExtraLineBreak")]
         public void CleanExtraLineBreak_HelloWorld_ReturnStringWithoutDoubleSLineBreak()
         {
             // Arrange
